Refactored the code, separating the game, experiment, player, and evaluation components.
