+++ k-reasoning/G08A/run.sh
@@ -1,0 +1,3 @@
+python main.py  --player_strategy "agent" --computer_strategy "fix" --max_round 10 --exp_num 1
+
+python evaluate.py --players agent --opponents fix
