+++ k-reasoning/SAG/run.sh
@@ -1,0 +1,3 @@
+python main.py  --player_strategy "kr" --computer_strategy "agent" --max_round 10 --exp_num 1
+
+python evaluate.py --players kr --opponents agent
